+++ tests/Traits/InteractsWithHttp.php
@@ -1,15 +1,10 @@
-<?php
-
-declare(strict_types=1);
+<?php declare(strict_types=1);
 
 /*
- * This file is part of Biurad opensource projects.
+ * This file is part of RadePHP Demo Project
  *
- * PHP version 8.0 and above required
- *
- * @author    Divine Niiquaye Ibok <divineibok@gmail.com>
- * @copyright 2019 Biurad Group (https://biurad.com/)
- * @license   https://opensource.org/licenses/BSD-3-Clause License
+ * @copyright 2022 Divine Niiquaye Ibok (https://divinenii.com/)
+ * @license   https://opensource.org/licenses/MIT License
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -19,7 +14,7 @@ namespace App\Tests\Traits;
 
 use Biurad\Http\ServerRequest;
 use Flight\Routing\Route;
-use Flight\Routing\Router;
+use Psr\Http\Message\StreamInterface;
 use Psr\Http\Message\UriInterface;
 
 trait InteractsWithHttp
@@ -35,29 +30,32 @@ trait InteractsWithHttp
         array $query = [],
         array $headers = [],
         array $cookies = []
-    ): Route {
+    ): ?Route {
         $request = $this->request($uri, $method, $query, $headers, $cookies);
 
         foreach ($attributes as $key => $value) {
             $request = $request->withAttribute($key, $value);
         }
 
-        return $this->makeApp()->get(Router::class)->matchRequest($request);
+        return $this->makeApp()->getRouter()->matchRequest($request);
     }
 
     /**
      * This helper method abstracts the boilerplate code needed to test the
      * execution of a server request.
+     *
+     * @param resource|StreamInterface|string|null $body
      */
     public function request(
         string|UriInterface $uri,
         string $method = 'GET',
+        $body = 'php://input',
         array $query = [],
         array $headers = [],
         array $cookies = []
     ): ServerRequest {
         $servers = $_SERVER + ['SERVER_PROTOCOL' => 'HTTP/1.1'];
 
-        return (new ServerRequest($method, $uri, $headers, 'php://input', '1.1', $servers))->withCookieParams($cookies)->withQueryParams($query);
+        return (new ServerRequest($method, $uri, $headers, $body, '1.1', $servers))->withCookieParams($cookies)->withQueryParams($query);
     }
 }